+++ packages/stream_video_flutter/lib/src/call_controls/controls/toggle_recording_option.dart
@@ -11,6 +11,10 @@ class ToggleRecordingOption extends StatelessWidget {
     required this.call,
     this.enabledRecordingIcon = Icons.video_file_outlined,
     this.disabledRecordingIcon = Icons.video_file_rounded,
+    this.enabledRecordingIconColor,
+    this.disabledRecordingIconColor,
+    this.enabledRecordingBackgroundColor,
+    this.disabledRecordingBackgroundColor,
   });
 
   /// Represents a call.
@@ -22,12 +26,29 @@ class ToggleRecordingOption extends StatelessWidget {
   /// The icon that is shown when the recording is disabled.
   final IconData disabledRecordingIcon;
 
+  /// Color of the icon when recording is enabled
+  final Color? enabledRecordingIconColor;
+
+  /// Color of the icon when recording is disabled
+  final Color? disabledRecordingIconColor;
+
+  /// Color of the background when recording is enabled
+  final Color? enabledRecordingBackgroundColor;
+
+  /// Color of the background when recording is disabled
+  final Color? disabledRecordingBackgroundColor;
+
   @override
   Widget build(BuildContext context) {
     final enabled = call.state.value.isRecording;
 
     return CallControlOption(
       icon: enabled ? Icon(enabledRecordingIcon) : Icon(disabledRecordingIcon),
+      iconColor:
+          enabled ? enabledRecordingIconColor : disabledRecordingIconColor,
+      backgroundColor: enabled
+          ? enabledRecordingBackgroundColor
+          : disabledRecordingBackgroundColor,
       onPressed: () {
         if (!enabled) {
           call.startRecording();