+++ packages/stream_video_flutter/lib/src/call_participants/regular_call_participants_content.dart
@@ -53,6 +53,8 @@ class RegularCallParticipantsContent extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
+    final participantsTheme = StreamCallParticipantTheme.of(context);
+
     final remoteParticipants = participants.where((e) => !e.isLocal);
     final localParticipant = participants.where((e) => e.isLocal).first;
 
@@ -87,6 +89,9 @@ class RegularCallParticipantsContent extends StatelessWidget {
       call: call,
       participants: gridParticipants,
       itemBuilder: callParticipantBuilder,
+      padding: participantsTheme.participantsGridPadding,
+      mainAxisSpacing: participantsTheme.participantsGridMainAxisSpacing,
+      crossAxisSpacing: participantsTheme.participantsGridCrossAxisSpacing,
     );
 
     if (showLocalVideo) {