+++ packages/stream_video_flutter/lib/src/call_screen/call_content/call_app_bar.dart
@@ -19,9 +19,6 @@ class CallAppBar extends StatelessWidget implements PreferredSizeWidget {
     this.elevation = 1,
     this.backgroundColor,
     this.onBackPressed,
-    this.onLayoutModeChanged,
-    this.onParticipantsInfoTap,
-    this.participantsInfoBuilder,
     this.leading,
     this.title,
     this.actions,
@@ -41,14 +38,6 @@ class CallAppBar extends StatelessWidget implements PreferredSizeWidget {
 
   /// The action to perform when the back button is pressed.
   final VoidCallback? onBackPressed;
-
-  /// The action to perform when the participants info button is tapped.
-  final VoidCallback? onParticipantsInfoTap;
-
-  final ValueSetter<ParticipantLayoutMode>? onLayoutModeChanged;
-
-  /// Builder used to create a custom participants info screen.
-  final CallParticipantsInfoBuilder? participantsInfoBuilder;
 
   /// The leading widget to display.
   final Widget? leading;
@@ -86,20 +75,12 @@ class CallAppBar extends StatelessWidget implements PreferredSizeWidget {
     return AppBar(
       elevation: elevation,
       leading: leadingWidget,
+      leadingWidth: 120,
       backgroundColor: backgroundColor ?? theme.colorTheme.barsBg,
-      centerTitle: false,
+      centerTitle: true,
       actions: actions ??
           <Widget>[
-            ParticipantLayoutModeButton(
-              onLayoutModeChanged: onLayoutModeChanged,
-            ),
-            IconButton(
-              icon: Icon(
-                Icons.group_rounded,
-                color: theme.colorTheme.textHighEmphasis,
-              ),
-              onPressed: () => _onParticipantsInfoTap(context),
-            ),
+            LeaveCallOption(call: call),
           ],
       title: title ??
           Text(
@@ -114,63 +95,6 @@ class CallAppBar extends StatelessWidget implements PreferredSizeWidget {
     final callState = call.state.value;
 
     final status = callState.status.toStatusString();
-    final callId = callState.callCid.id;
-    return '$status: $callId';
-  }
-
-  void _onParticipantsInfoTap(BuildContext context) {
-    if (onParticipantsInfoTap != null) {
-      onParticipantsInfoTap!();
-    } else {
-      final usersProvider = StreamUsersConfiguration.of(context);
-
-      Navigator.of(context).push(
-        MaterialPageRoute<Widget>(
-          builder: (context) =>
-              participantsInfoBuilder?.call(context, call) ??
-              StreamCallParticipantsInfoMenu(
-                call: call,
-                usersProvider: usersProvider,
-              ),
-        ),
-      );
-    }
-  }
-}
-
-class ParticipantLayoutModeButton extends StatefulWidget {
-  const ParticipantLayoutModeButton({
-    super.key,
-    this.onLayoutModeChanged,
-  });
-
-  final ValueSetter<ParticipantLayoutMode>? onLayoutModeChanged;
-
-  @override
-  State<ParticipantLayoutModeButton> createState() =>
-      _ParticipantLayoutModeButtonState();
-}
-
-class _ParticipantLayoutModeButtonState
-    extends State<ParticipantLayoutModeButton> {
-  ParticipantLayoutMode _layoutMode = ParticipantLayoutMode.grid;
-
-  @override
-  Widget build(BuildContext context) {
-    return IconButton(
-      icon: Icon(
-        _layoutMode == ParticipantLayoutMode.grid
-            ? Icons.highlight_rounded
-            : Icons.grid_view_rounded,
-      ),
-      onPressed: () {
-        _layoutMode = _layoutMode == ParticipantLayoutMode.grid
-            ? ParticipantLayoutMode.spotlight
-            : ParticipantLayoutMode.grid;
-
-        widget.onLayoutModeChanged?.call(_layoutMode);
-        setState(() {});
-      },
-    );
+    return status;
   }
 }